Add restricted picklist option
When creating a picklist field using `force field create My_Widget__c "Widget Type":picklist picklist:"Fancy, Standard"` it is not possible to specify if the picklist should be restricted to the given values. The Force CLI returns "ERROR: validation error: restrictedPicklist:false is not a valid option for field type picklist".

This PR would add the ability to specify `restrictedPicklist:false` when creating a new picklist field.